LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/geolocation-API/disabled-by-permissions-policy.https.sub-expected.txt

CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/resources/testdriver.js from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/resources/testdriver-vendor.js from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Permission policy 'Geolocation' check failed for document with origin 'https://127.0.0.1:9443'.


FAIL Permissions-Policy header geolocation=() disallows the top-level document. promise_test: Unhandled rejection with value: object "[object GeolocationPosition]"
FAIL Permissions-Policy header geolocation=() disallows same-origin iframes. assert_false: Geolocation API expected false got true
FAIL Permissions-Policy header geolocation=() disallows cross-origin iframes. assert_false: Geolocation API expected false got true
